Fix Go To Definition and also make it search "All Projects"
I am currently still working on this, hence why I just commented things
out/etc.  However in summary, it appears that the current command used
for "Goto Definition" was no longer working with Eclim 1.7.4.  This
change basically sends a pattern instead of a file/word offset.  It also
will select the first result instead of only working if one result
exsits.  I plan to update this to show a file list as well.  I will also
cleanup this code change in the future.