List choices for "Go to Definition"  / Account for line endings
Sublime does not report line ending characters in its offset counts
which casues eclim to return completion results for the wrong objects.
I currently not account for line endings by adding a character per line.
I plan to make this a configurable value incase of multi-character line
endings.  I am not sure if this is necessary, but who knows.